FEAT: GO Converted to API
converted the Go code to serve as a web API
which write out JSON results in order to prepare
for integration to future front end. Currently,
the API will run a Local Host server on port
8080 which will display the created JSON
elements however in future this information
will be taken directly into the front end
of the website to be displayed in a more
graphically pleasing manner.