Feat: Created Base Meal Slots
Created a basic meal slot selection system
which allows user to "spin" for a simple meal
comprised of a protein, a vegetable, and a starch.
While refinements will be made, this system is in
a working stage, though not quite finished.
Further refinements will include refinements to
the overall look and feel of the machine,
refinement of the overall reels, and work on the
overall functionality of the wheel.