FEAT: Inserted Recipes into Dessert Table
Inserted 37 recipes into the Desserts Table in the
dessert recipe sql server. This will later be used
to go through the Golang API to be served to the
desserts front end segment. As well, an Admin
Interface will be coded in to make the process
of adding desserts to the database easier in the
future.