FEAT: Desserts Showcase Completion
Completed Desserts Showcase utilizing a Golang
backend API to serve up a list of desserts which
are then interpreted and used to show off a
polaroid style of baked goods that can be made
to pair with any of the dinners from the meal slot
machine. Gallery of desserts then offers the needed
ingredients, as well as the approximate bake time
and the link to the original recipe to follow
the instructions and make the signature sweet of
users choosing. This is accomplished through a
modal menu in JavaScript which lists out these elements
from the SQL server which is brought in through as
a JSON file through the API, and stored in an Array.
This allows each element to be called and used as
needed to complete the desired task.