FEAT: Admin Window Application Created
Created an Admin Window for updating the Meals.db
file in order to ease the work of assigning recipe
and image links for future updating in order to
get the backend ready for the next stage of
site development. This application utilizes a
tkinter interface to create a window which displays
all the elements currently in the Meals.db database
and allow for the user to add, edit, or delete Meals
as needed. This will only edit the persons local
version of Meals.db, and will need to be changed
to a more restricted access if deployed further.